requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr==1.7.0
transformers==4.35.2
torch==2.5.0
pydantic==2.5.0
openai==1.3.7
twilio==8.10.0
schedule==1.2.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
